Added map_conf1
If write the address and size in host0:map_conf/map_conf1 with Little Edian, the contents will be reflected in host0:map_conf/map_mem1.
The address must be that of the kernel.
If it collides with an unmapped page, the device will trigger a kernel panic and reboot, or you will see a bsod.